Create #6
A school camp is organized by a school to support the process of preparing their students for
an examination. They are in need of a study timetable that has following assumptions:
Assumptions:
1. Total Days of Camp – 5 Days
2. Total Hours a day – 5 Days
3. Total Subjects – 5 Subjects
Note: The timetable should not follow the same order and should be in random everyday.
Prepare code logic to help the School.